pyproject.toml
[tool.poetry]
name = "investments"
version = "0.1.28"
description = "Analysis of Interactive Brokers reports for tax reporting in Russia"
license = "MIT"
authors = ["Maxim Andreev <andreevmaxim@gmail.com>"]
readme = "README.md"
repository = "https://github.com/qk4l/investments"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Office/Business :: Financial :: Accounting",
    "Topic :: Office/Business :: Financial :: Investment"
]

[tool.poetry.scripts]
ibtax = "investments.ibtax.ibtax:main"
ibdds = "investments.ibdds.ibdds:main"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.5"
requests = "^2.31"
tabulate = "^0.9"
aiomoex = "^2.0"
WeasyPrint = "^58.0"
jinja2 = "^3.1"
google-auth-oauthlib = "1.1.0"
google_spreadsheet = "0.0.6"
googleapis-common-protos = "1.61.0"


[tool.poetry.dev-dependencies]
pytest = "^7.2"
mypy = "^1.1"
wemake-python-styleguide = "^0.17"
types-requests = "^2.28"
types-tabulate = "^0.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
